homework, added other tables/fields to project

+++ Module305JDBC/Module305JDBC/src/main/java/org/example/database/entity/Product.java
@@ -19,22 +19,32 @@ public class Product {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
     private Integer id;
+
     @Column(name = "product_code")
     private String productCode;
+
     @Column(name = "product_name")
     private String productName;
+
     @Column(name = "productline_id")
     private Integer productlineId;
+
     @Column(name = "product_scale")
     private String productScale;
+
     @Column(name = "product_vendor")
     private String productVendor;
+
     @Column(name = "product_description")
     private String productDescription;
+
     @Column(name = "quantity_in_stock")
-    private Short quantityInStock;
+    private Integer quantityInStock;
+
     @Column(name = "buy_price")
     private BigDecimal buyPrice;
+
     @Column(name = "msrp")
     private BigDecimal msrp;
 }
+
